fix: optimize port types in MSR module
- Updated `ServiceEndpoint` and `ServiceInstance` to use `i32` for `port`, aligning with `Microservice` and PostgreSQL `INTEGER`.
- Removed redundant `as i32` cast in `db.rs`.
- Verified API Gateway compilation.

+++ crates/msr/src/db.rs
@@ -31,7 +31,7 @@ impl MicroserviceRepository {
         .bind(&service.name)
         .bind(&service.version)
         .bind(&service.host)
-        .bind(service.port as i32)
+        .bind(service.port)
         .bind(&service.protocol)
         .bind(service.status)
         .bind(sqlx::types::Json(&service.metadata))
