+++ crates/msr/src/service.rs
@@ -37,7 +37,7 @@ pub enum ServiceStatus {
 pub struct ServiceEndpoint {
     pub service_id: Uuid,
     pub host: String,
-    pub port: u16,
+    pub port: i32,
     pub protocol: String,
     pub weight: u32,
 }
@@ -48,7 +48,7 @@ pub struct ServiceInstance {
     pub id: Uuid,
     pub service_id: Uuid,
     pub host: String,
-    pub port: u16,
+    pub port: i32,
     pub status: ServiceStatus,
     pub metadata: HashMap<String, String>,
     pub registered_at: chrono::DateTime<chrono::Utc>,
@@ -100,7 +100,7 @@ impl Microservice {
 
 impl ServiceEndpoint {
     /// Create new service endpoint
-    pub fn new(service_id: Uuid, host: String, port: u16) -> Self {
+    pub fn new(service_id: Uuid, host: String, port: i32) -> Self {
         Self {
             service_id,
             host,
@@ -118,7 +118,7 @@ impl ServiceEndpoint {
 
 impl ServiceInstance {
     /// Create new service instance
-    pub fn new(service_id: Uuid, host: String, port: u16) -> Self {
+    pub fn new(service_id: Uuid, host: String, port: i32) -> Self {
         Self {
             id: Uuid::new_v4(),
             service_id,
